requirements.txt
streamlit==1.28.0
vosk==0.3.45
sounddevice==0.4.6
opencv-python==4.8.1.78
gTTS==2.3.2
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy==1.24.3
pillow==10.0.1
pygame==2.5.2
pydub==0.25.1
speechrecognition==3.10.0